fixed jobs screen

+++ lib/electrical_components/transformer_trainer/modes/guided_mode.dart
@@ -336,7 +336,7 @@ class GuidedModeWidget extends StatelessWidget {
         child: AlertDialog(
           backgroundColor: PopupThemeData.alertDialog().backgroundColor,
           shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(PopupThemeData.alertDialog().borderRadius),
+            borderRadius: BorderRadius.circular(PopupThemeData.alertDialog().borderRadius as double),
             side: BorderSide(
               color: PopupThemeData.alertDialog().borderColor,
               width: PopupThemeData.alertDialog().borderWidth,
