+++ lib/electrical_components/transformer_trainer/widgets/trainer_widget.dart
@@ -220,7 +220,7 @@ class TransformerTrainer extends StatelessWidget {
         child: AlertDialog(
           backgroundColor: PopupThemeData.alertDialog().backgroundColor,
           shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(PopupThemeData.alertDialog().borderRadius),
+            borderRadius: BorderRadius.circular(PopupThemeData.alertDialog().borderRadius as double),
             side: BorderSide(
               color: PopupThemeData.alertDialog().borderColor,
               width: PopupThemeData.alertDialog().borderWidth,
